Null Ptr Check
In some cases there could be a SECFailure when using nss_cmd but no
actual PR_Error, in that case calling std::str::raw::from_c_str on
PR_ErrorToName(PR_GetError()) could result in a null ptr trying to be
dereferenced. So Split up the command and check for ptr::null().

As a bonus it's probably slightly easier to read.